.requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
cachetools==5.2.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
cycler==0.11.0
fonttools==4.33.3
frozenlist==1.3.0
fsspec==2022.5.0
google-auth==2.6.6
google-auth-oauthlib==0.4.6
grpcio==1.46.3
idna==3.3
kiwisolver==1.4.2
Markdown==3.3.7
matplotlib==3.5.2
multidict==6.0.2
numpy==1.22.4
oauthlib==3.2.0
packaging==21.3
pandas==1.4.2
Pillow==9.1.1
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyDeprecate==0.3.2
pynvml==11.4.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytorch-lightning==1.6.3
pytz==2022.1
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scipy==1.8.1
six==1.16.0
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
torch-tb-profiler==0.4.0
torchmetrics==0.9.0
tqdm==4.64.0
typing_extensions==4.2.0
urllib3==1.26.9
Werkzeug==2.1.2
yarl==1.7.2
